+++ docs/changelog.md
@@ -6,6 +6,7 @@
 
 + UI adjustments.
 + Correction of zoom value.
++ Add "default zoom value" option.
 
 **Development**: